refactor(app): ConflictTextEdit 抽為獨立檔案
ConflictTextEdit(帶行號欄的 QPlainTextEdit)即將為了直接操作(hover 浮起、
拖曳採用、點行組合)長出不少狀態與事件處理,留在 ConflictResolvePanel.cpp
裡會讓這個檔案遠超過本專案 800 行的規範上限。純搬移,零行為變更。

- 新增 src/app/views/ConflictTextEdit.{h,cpp},ConflictTextEdit 與
  LineNumberArea 從 ConflictResolvePanel.cpp 的匿名 namespace 原封不動搬過去。
- LineNumberArea 改放在具名的 gbm namespace(而非匿名 namespace):
  ConflictTextEdit.h 裡 `friend class LineNumberArea;` 是不具名字首的
  friend 宣告,會具體指向 gbm::LineNumberArea;若 LineNumberArea 留在匿名
  namespace,會變成另一個同名但不相干的類別,friend 關係對不上,私有成員
  就存取不到 —— 這是搬移過程中發現、必須一併處理的細節,不是額外功能。
- ConflictResolvePanel.cpp 移除三個因此不再需要的 include(QPainter/
  QResizeEvent/QTextBlock)。
- src/app/CMakeLists.txt 與 gbm_conflict_ui_tests 的來源清單都加入新檔。

全部 289 項單元測試(含 gbm_conflict_ui_tests)維持 100% 通過,行為不變。
